refactor: PositiveNumber 초기값을 양수만 받을 수 있도록 수정
